Wrote the initial example program.
Several things to note:
o The whole API needs to be improved.  We need to make it easier to
  promote pawns, e.g.
o The AI code needs a cleaner way of checking for pawn promotion, and it
  needs to have a cleaner feel all over.
o Refactor/clean-up the example code to be nicer to look at.
o Learn from the example code, and fix the API where it feels difficult
  to work with.