Bug in the updating of the pawn bitboards.
I've written a new unit test for it in check_board.c that will fail.  To
reproduce, run the following:
1. make examples
2. gdb ./a.out
3. break fc_board_make_move
4. run 1 3
5. a4-a5
6. Now step through the updating of the pawn orientation bitboards.
Right before it executes board.c:655, the program suddenly returns from
the function, leaving the bitboards incompletely updated.  This problem
is present with at least -O0, -O2, and -O3.

Try to install a different version of gcc, e.g. 6, and see if the
problem is still present.