Init'ed all the mlists before calling alphabeta.
This puts all the initialization of the mlists for the AI before we call
alphabeta so that we aren't callocing and freeing all throughout the AI
process.

I had hoped that it would result in a speed-up, but (as gprof predicted)
it didn't.  I don't know why this should be.  I really should trust
gprof more than myself.