Added bitboards to keep track of threats.
This strategy turns out to take MORE time than just checking for threats
on the fly.  There are a couple of ideas I have for using the threats.c
file as is:

(1) Figure out how to generate the mlists from the threat bitboards.
	- I don't see how this would speed anything up.
(2) Generate the threat bitboards as we generate the moves; keep these
	bitboards in the move structure.  Then when we make_move(), we
	can just memcpy() the bitboards from the move struct onto the
	game board.
	- This seems overly convoluted.

...or...

(3) Suck it up and use setjmp and longjmp to handle kings getting
captured.
	- This will be VERY convoluted, but highly likely to be faster
	  than any of the above options.

We need to speed up the AI by a factor of 10 in order for it to be
useful at all.