Removes redundant move validation code from AI.
This commit fully moves all the game logic from the AI source to the
board source.  A couple moves are taking longer than expected (3Xs
longer), but over-all I think this is a good decision, and it will make
it easier to both add the parallel moves and negascout to the AI code.