Fleshes out the threaded move search algorithm.
Updates #7.

There are a couple of things that bug me about the (now working)
threaded search:

First, it doesn't seem to be any faster (for 2 threads).  Of course, it
may turn out that multi-threaded search is only useful on 4 or 8 core
machines -- needs more testing.

Second, the depth adjustments that I need to make in the threaded
algorithm are THE EXACT OPPOSITE of those in the regular alphabeta
algorithm.  I'm not sure what's going on there.

Other than the above troubling items, I need to go through all the added
code and refactor it before I call this issue closed.

Also, future topic:  We need to make the fc_board_get_valid_moves()
function take as close to 0 time as we can get.  As is, it is slower
(but not by much) than the old way when I was baking in the game logic
into alphabeta.