Fixes failing test from prior commit (af66db0b96).
Let me tell you a story about the strangest bug I ever had to fix:

It started with tests failing sometimes due to segfaults and sometimes
floating point exceptions.  And even occationally, the test would pass
without a hitch.  When I initially ran the code in GDB, assertions would
fail in the next test case (AI) but no errors were generated.
Eventually I dinked around in GDB long enough trying to figure out what
was wrong that I saw a line printed to the terminal from the AI tests,
and immediately after a segfault in one of the threads from the thread
tests which, I thought, should have already finished.

Ah ha! I thought.  I'm calling pthread_cancel to end the threads, but it
looks like I can't trust it.  The threads are continuing to run beyond
their expected lifetime, and the stack (and by extension, the structures
they depended on) were changing out from under them.  Sometimes this
meant that the thread would sefault or that I would get a floating point
exception (from the modulus operator failing for some reason), and
sometimes this meant that the boards on the AI stack were being changed
underneath the AI which caused the assertions being thrown.

So I changed the pthread_cancel calls to pthread_kill calls and created
signal handlers for all the threads.  When the thread was sent a
SIGUSR2, the handler would be invoked which simply called pthread_exit.
Maybe I could trust this implementation, I thought.  And indeed, the
first few tests ran fine, but the seventh or eighth time I ran 'make
clean check', the test hung.  Check wasn't very helpful at telling me
exactly where, and since I didn't know if something in one of the child
threads was failing or if it was the main thread, I decided to sprinkle
printfs and fflushes throughout and narrow the hanging line down.

I finally discovered that the pthread_mutex_destroy call in
fc_tpool_free was what was hanging.  Doing some research I discovered
that calling pthread_mutex_destroy on a locked mutex resulted in
undefined behavior.  What I figure was happening was that a thread was
locking the mutex and then getting the signal interrupt.  When
pthread_mutex_destroy was then called on the locked thread, it hung.
Perhaps because the Mac's implementation of pthread_mutex_destroy tries
to acquire the lock before freeing it; I can't say.  Pthread_cancel must
have been handling some clean-up of the threads/mutexes before exiting.

So I reimplemented fc_tpool_kill_threads with pthread_cancel and put a
sleep(2) in the failing/erroring test case to give the test threads
enough time to quit before calling the next test.  I understand that
this is a hack; however, fc_tpool_kill_threads is not essential to
libforchess at this time, so I can afford it right now.  Plus I intend
to refactor the thread code next based on what I'm reading in O'Reilly's
Pthreads Programming and Principles of Concurrent and Distributed
Programming by M. Ben-Ari, so a great deal of the implmentation will
probably change anyway.

Morals of the story:
(1) Read the manual.  Even if a function call seems self-explanatory
    like pthread_mutex_destroy, you should still bother to check the man
    page for any caveats, known bugs, or undefined behavior.
(2) Undefined behavior means exactly that.  You can't trust it nor can
    you fully anticipate what it will do.  You may think that it just
    means that your program may crash, but it could instead hang
    unpredictably.  It could also download child porn to your computer
    and anonymously notify the authorities.  You just can't know for
    sure, so it's best to avoid it altogether.

THE END

P.S. What made these bugs so difficult to track down was that I started
trying to find them by commenting out certain lines of the code (such as
fc_tpool_kill_threads) which did not result in a hung test and initially
made me think that something in fc_tpool_kill_threads was hanging
instead of fc_tpool_free.